better

+++ std/thr/latch.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <std/sys/types.h>
+
 namespace stl {
     class Latch {
         struct Impl;
         Impl* impl;
 
     public:
-        explicit Latch(int n);
+        explicit Latch(size_t n);
         ~Latch() noexcept;
 
         void arrive() noexcept;
